Read the description of this commit to understand these changes.
This is a shit ton of changes.

First of all full markdown support has been implemented and is working. The preview can be found here: https://imgur.com/a/qLMj2zb (It is still a work-in-progress and is buggy!)

Second, message loading has been implemented and is working great. I'll pair it up with the Websockets in the future.

Third, there is now a placeholder for when a DM isn't opened that says "Open a DM / server to get started!" which is a nice touch.

Websockets and server loading is still being implemented and is a work-in-progress.

Read the functionality list to get an understanding of what's implemented.