Read the description of this commit to understand these changes.
Hey! This was also a shit ton of changes.

Websockets are now done! We can now receive messages instantly. I will add more functionality to them later on in development.

You can now send messages! Although uploading files hasn't been implemented yet, but that is nice

Third, you can see users statuses! This was a change I wanted to add for a long time but now they work and they work great. This does mean we need to add a delay before loading the UI (1.5 second delay) but I have added a loading screen for fun.

The markdown rendering (MDLabel) has been partially rewritten for it to render properly. For example, especially visible when using italic there were spacing issues and those have now been fixed.

The API has also gotten more debugging so I can see where issues are in the API.

We are close to a proper release to this new era of Naticord.