Read the description of this commit to understand these changes.
A lot of new changes, again.

A new logo! This is a bit more unique from the older logo and I like how it turned out. This was also liked by the community.

Server support is getting there! Servers now load into the list and you can click on them, although they do nothing. I'm still thinking on how to implement the loading the categories.

The MDLabel now has text wrapping and has gotten code improvements! This makes it more stable and makes it also look nicer!

Some extra functions got some changes due to servers but it's mostly the same.

A bug was fixed that when you login with 2FA the main login window wouldn't close, now it does.

Beta 2 is on the horizon!