A huge update! Read the description of this commit
This has been a huge update, it's a still a work in progress but I'm very much liking it so far.

The client UI has been revamped entirely with an extended titlebar with options that has Aero, Acrylic, Mica (Alt) backdrop capabilities (thanks to aerocord) and it's still not done, but it's looking great so far.

OS Version grabbing has also been introduced (to help with backdrop fetching) and it works great.

Classic mode is still getting fixes, but in the final release classic, basic and composition should all look great.

So far, so good.